requirements.txt
google-generativeai>=0.3.0
googlemaps>=4.10.0
python-dotenv>=1.0.0